tests/ci: style: canonicalize, sort, and multiline kmods/pkgs
This patch helps to remove ambiguity and reduce future diffs of changes
to the VM_EXTRA_PACKAGES list. This patch also sorts kld_list.

NOTE: lang/perl5 does not exist, so we must use the perl5 pkgname
in order to always track the latest version in DEFAULT_VERSIONS.

Signed-off-by: Siva Mahadevan <me@svmhdvn.name>
Sponsored by: The FreeBSD Foundation